requirements.txt
# Python version: 3.10.11


# 核心框架與網頁組件
dash==4.0.0
dash-bootstrap-components==2.0.4
dash-core-components==2.0.0
dash-leaflet==1.0.15                   # 地圖顯示功能
dash-table==5.0.0
geopandas==1.1.3                      # 地圖資料
selenium==4.41.0


# 數據處理與機器學習
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2                      # (注意安裝名稱)



# 環境配置
python-dotenv==1.2.2               # 讀取 .env 設定檔
